requirements.txt
# BOX MAGIC OCR INTELLIGENT - Requirements
# Version : 1.0.0

# Core dependencies
pyyaml>=6.0

# PDF processing (choose one or more)
PyPDF2>=3.0.0              # For text-based PDFs
pdfplumber>=0.10.0         # Alternative PDF text extraction

# OCR for scanned documents (optional but recommended)
pytesseract>=0.3.10        # Tesseract wrapper
Pillow>=10.0.0             # Image processing
pdf2image>=1.16.0          # Convert PDF to images for OCR

# Google Sheets integration (optional)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development (optional)
black>=23.0.0              # Code formatting
flake8>=6.0.0              # Linting
mypy>=1.5.0                # Type checking
